Fix build

+++ include/game/board/basic_space.h
@@ -3,8 +3,8 @@
 
 #include "dolphin.h"
 
-void BoardEventLandBlue(s32, s16);
-void BoardEventLandRed(s32, s16);
+void BoardEventLandBlue(s32, s32);
+void BoardEventLandRed(s32, s32);
 s8 BoardCoinChgCreate(Vec*, s8);
 s32 BoardCoinChgExist(s32);
 void BoardCoinChgHide(s32);
